+++ src/tree/database/CollectionTreeItem.ts
@@ -1,13 +1,13 @@
 import { ThemeIcon, TreeItem, TreeItemCollapsibleState } from "vscode";
 import { Collection } from "../../appwrite";
 import { databaseClient } from "../../client";
-import { AwTreeItem } from "../common/AwTreeItem";
+import { AppwriteTreeItemBase } from "../../ui/AppwriteTreeItemBase";
 import { DatabaseTreeItemProvider } from "./DatabaseTreeItemProvider";
 import { DocumentsTreeItem } from "./DocumentsTreeItem";
 import { PermissionsTreeItem } from "./settings/PermissionsTreeItem";
 import { RulesTreeItem } from "./settings/RulesTreeItem";
 
-export class CollectionTreeItem extends AwTreeItem {
+export class CollectionTreeItem extends AppwriteTreeItemBase {
     constructor(public collection: Collection, public readonly provider: DatabaseTreeItemProvider) {
         super(undefined, collection.name);
     }