+++ src/tree/database/DatabaseTreeItemProvider.ts
@@ -5,7 +5,7 @@ import { Collection, CollectionsList } from "../../appwrite";
 import { CollectionTreeItem } from "./CollectionTreeItem";
 import { AppwriteSDK } from "../../constants";
 import { ext } from '../../extensionVariables';
-import { AwParentTreeItem } from '../common/AwParentTreeItem';
+import { AppwriteTreeItemBase } from '../../ui/AppwriteTreeItemBase';
 
 export class DatabaseTreeItemProvider implements vscode.TreeDataProvider<vscode.TreeItem> {
     private _onDidChangeTreeData: vscode.EventEmitter<vscode.TreeItem | undefined | void> = new vscode.EventEmitter<
@@ -33,8 +33,8 @@ export class DatabaseTreeItemProvider implements vscode.TreeDataProvider<vscode.
             return Promise.resolve([]);
         }
 
-        if (parent instanceof AwParentTreeItem) {
-            return parent.getChildrenInternal() ?? [];
+        if (parent instanceof AppwriteTreeItemBase) {
+            return await parent.getChildren?.() ?? [];
         }
 
         const databaseSdk = new AppwriteSDK.Database(client);