+++ src/tree/functions/FunctionTreeItem.ts
@@ -1,13 +1,13 @@
 import { MarkdownString, ThemeIcon, TreeItem, TreeItemCollapsibleState } from "vscode";
 import { Function } from "../../appwrite";
-import { AwTreeItem } from "../common/AwTreeItem";
+import { AppwriteTreeItemBase } from "../../ui/AppwriteTreeItemBase";
 import { msToDate } from '../../utils/date';
 import { ExecutionsTreeItem } from './executions/ExecutionsTreeItem';
 import { FunctionsTreeItemProvider } from './FunctionsTreeItemProvider';
 import { FunctionSettingsTreeItem } from './settings/FunctionSettingsTreeItem';
 import { TagsTreeItem } from './tags/TagsTreeItem';
 
-export class FunctionTreeItem extends AwTreeItem {
+export class FunctionTreeItem extends AppwriteTreeItemBase {
     constructor(public func: Function, public readonly provider: FunctionsTreeItemProvider) {
         super(undefined, func.name);
         this.tooltip = new MarkdownString(`ID: ${func.$id}  \nLast updated: ${msToDate(func.dateUpdated)}  \nCreated: ${msToDate(func.dateCreated)}`);