+++ src/tree/functions/FunctionsTreeItemProvider.ts
@@ -2,11 +2,10 @@ import * as vscode from "vscode";
 import { client } from "../../client";
 import { Function, FunctionsList } from "../../appwrite";
 import { AppwriteSDK } from "../../constants";
-import { AwTreeItem } from "../common/AwTreeItem";
+import { AppwriteTreeItemBase } from "../../ui/AppwriteTreeItemBase";
 import { ext } from "../../extensionVariables";
 import { EventEmitter, TreeItem } from "vscode";
 import { FunctionTreeItem } from "./FunctionTreeItem";
-import { AwParentTreeItem } from '../common/AwParentTreeItem';
 
 export class FunctionsTreeItemProvider implements vscode.TreeDataProvider<vscode.TreeItem> {
     private _onDidChangeTreeData: EventEmitter<TreeItem | undefined | void> = new EventEmitter<TreeItem | undefined | void>();
@@ -26,7 +25,7 @@ export class FunctionsTreeItemProvider implements vscode.TreeDataProvider<vscode
         return element;
     }
 
-    async getChildren(parent?: AwTreeItem | TreeItem): Promise<vscode.TreeItem[]> {
+    async getChildren(parent?: AppwriteTreeItemBase | TreeItem): Promise<vscode.TreeItem[]> {
         if (client === undefined) {
             return Promise.resolve([]);
         }
@@ -44,8 +43,8 @@ export class FunctionsTreeItemProvider implements vscode.TreeDataProvider<vscode
             return [{ label: "No functions found" }];
         }
 
-        if (parent instanceof AwParentTreeItem) {
-            return parent.getChildren();
+        if (parent instanceof AppwriteTreeItemBase) {
+            return await parent.getChildren?.() ?? [];
         }
 
         return [];