+++ src/tree/functions/executions/ExecutionTreeItem.ts
@@ -1,9 +1,6 @@
 import { MarkdownString, ThemeColor, ThemeIcon, TreeItem } from "vscode";
 import { Execution, ExecutionStatus } from "../../../appwrite";
-import { functionsClient } from "../../../client";
-import { ext } from "../../../extensionVariables";
 import { msToDate } from "../../../utils/date";
-import { sleep } from '../../../utils/sleep';
 import { ExecutionsTreeItem } from "./ExecutionsTreeItem";
 
 const executionStatusIcons: Record<ExecutionStatus, ThemeIcon> = {
@@ -26,33 +23,33 @@ export class ExecutionTreeItem extends TreeItem {
         this.description = execution.trigger;
         this.contextValue = this.getContextValue(execution);
         this.isAutoRefreshing = execution.status === "processing" || execution.status === "waiting";
-        if (this.isAutoRefreshing) {
-            this.autoRefresh();
-        }
+        // if (this.isAutoRefreshing) {
+        //     this.autoRefresh();
+        // }
     }
 
-    async autoRefresh(): Promise<void> {
-        if (!this.isAutoRefreshing) {
-            return;
-        }
-        this.refreshCount++;
-        ext.outputChannel.appendLog("Refreshing execution.");
-        const execution = await functionsClient?.getExecution(this.parent.parent.func.$id, this.execution.$id);
+    // async autoRefresh(): Promise<void> {
+    //     if (!this.isAutoRefreshing && this.refreshCount < 5) {
+    //         return;
+    //     }
+    //     this.refreshCount++;
+    //     ext.outputChannel.appendLog("Refreshing execution.");
+    //     const execution = await functionsClient?.getExecution(this.parent.parent.func.$id, this.execution.$id);
 
-        if (!execution) {
-            ext.outputChannel.appendLog("Execution is undefined");
-            this.isAutoRefreshing = false;
-            return;
-        }
-        this.execution = execution;
-        this.contextValue = this.getContextValue(execution);
-        this.iconPath = executionStatusIcons[execution.status];
-        this.label = this.getLabel(execution);
-        this.isAutoRefreshing = execution.status === "processing" || execution.status === "waiting";
-        ext.tree?.functions?.refreshChild(this);
-        await sleep(1000);
-        this.autoRefresh();
-    }
+    //     if (!execution) {
+    //         ext.outputChannel.appendLog("Execution is undefined");
+    //         this.isAutoRefreshing = false;
+    //         return;
+    //     }
+    //     this.execution = execution;
+    //     this.contextValue = this.getContextValue(execution);
+    //     this.iconPath = executionStatusIcons[execution.status];
+    //     this.label = this.getLabel(execution);
+    //     this.isAutoRefreshing = execution.status === "processing" || execution.status === "waiting";
+    //     ext.tree?.functions?.refreshChild(this);
+    //     await sleep(1000);
+    //     this.autoRefresh();
+    // }
 
     getLabel(execution: Execution): string {
         if (execution.status === "completed" || execution.status === "failed") {